Add support for importing a `fetchX` nix expression.
Nix is able to fetch and extract tarballs and git repositories in its
evaluation phase, which is used for fixing nixpkgs to a certain commit
in nix code. It provides `builtins.fetchTarball` and
`builtins.fetchGit` for that purpose.
We want to be able to re-use the resulting storepath as a bazel
target.

If you have a nix file like:

  builtins.fetchTarball {
    url = "example.com/mytarball.tar.gz";
    sha256 = …;
  }

You can make the contents of that tarball accessible as bazel targets
in the `WORKSPACE` file:

  nix_import_fetch_builtin(
    name = "mytarball",
    nix_file = "mytarball.nix",
  )

And have targets like:

  @mytarball//:file1
  @mytarball//:dir/file2

Requires `nix eval` (nix 2.0 or later), because `nix-instantiate` is
not able to print raw strings, but it should be possible to use
`nix-instantiate --eval` by unescaping the resulting string manually.