Fail when 'nix_file_deps' is not exhaustive
If `nix_file_deps` does not contain all the direct and indirect
dependencies of the nix package, bazel won't rebuild it if any of
these file change.

This patch track the dependencies by reading nix-build verbose output
and fails if theses dependencies are not explicitly listed in
nix_file_deps or repositories.