Simplify nix_file_deps checking
To check that dependencies of `nix_file_deps` are correctly declared,
we copy all declared dependencies into the external repository
directory. If any dependencies were missed, then nix evaluation will
fail.

To test that this really works, remove the

```
    nix_file_deps = ["//tests:pkgname.nix"],
```

line in `@nix-file-deps-test`. I tried writing a failure test for this
using Skylibs `analysistest` framework, but that doesn't seem to work
for failures in external repositories.

Like #76, this likewise closes #74, but in a simpler way.