Remove fail_not_supported attribute
This attribute was a hack, which is no longer used. It was meant to
allow hard dependencies on Nixpkgs packages even on platforms where
Nix does not exist (like Windows). But that is better solved through
other means, like tags that mark what targets are supported on what
platform. This flag was originally introduced for rules_haskell but is
no longer used there.