nixpkgs_python_configure: Use platform toolchain
This changes the Nix configured Python runtime from an "in-build
runtime" to a "platform runtime", see
https://docs.bazel.build/versions/2.0.0/be/python.html#py_runtime.files.

This is to avoid unecessary disk usage in the runfiles tree of targets
depending on the python toolchain.

In case of an "in-build runtime" Bazel sees all files that belong to the
toolchain, including the many include and Python module files. This
means that any target that has a runtime dependency on the Python
toolchain, will have each file contained in the Python toolchain listed
in its runfiles and runfiles manifest file. This adds around 4.6MiB to
the manifest file and around 34MiB to the runfiles tree. This space is
occupied repeatedly for each target depending on the nixpkgs python
toolchain.

E.g. on the rules_haskell repository:
```
$ du -ah bazel-bin/tests/run-tests@repl@repl.runfiles/{nixpkgs_python_toolchain_python3,MANIFEST}
...
4.7M    bazel-bin/tests/run-tests@repl@repl.runfiles/MANIFEST
34M     bazel-bin/tests/run-tests@repl@repl.runfiles/nixpkgs_python_toolchain_python3
```

After this change:
```
$ du -ah bazel-bin/tests/run-tests@repl@repl.runfiles/{nixpkgs_python_toolchain_python3,MANIFEST}
40K     bazel-bin/tests/run-tests@repl@repl.runfiles/MANIFEST
```