Invalid nixpkgs_package on nixpkgs_local_repository change
Additional Nix files of a `nixpkgs_local_repository` are not tracked
by the `nixpkgs_package` rule. In this commit, we generate the list of
these files in the `nixpkgs_local_repository` rule and make the
`nixpkgs_package` rule reading this list to add all of these files in
its own repository to let Bazel track them.

The list of files is propagated from one rule to the other one
throught a file named `nix-file-deps` created in the
`nixpkgs_local_repository` repository.

Closes #113