Test nixpkgs_package is invalidated on a change in Nix files
This test runs Bazel in the Bazel sandbox to ensure some dependencies
are well tracked at the workspace level. So, we create a workspace,
build it, then change the content of some workspace files and check if
a rebuild of this workspace takes into account these changes.

This is a regression test for the issue #113.