do not preserve generated README's permissions
this removes the requirement of having the generated file already
present in the source tree. otherwise, if it does not exist, `cp` will
fail with "Permission denied".

note that we had to explicitly make use of `POSIX_SH` provided by the
build itself, as `--no-preserve` is unique to `coreutils` and not
available on e.g. BSD or Darwin.