factor out dependencies for rendering module documentation
this is to have as little repetition as currently possible, and while it
looks ugly, it is intended to be only temporary until the
sub-repositories are proper `bzlmod` modules.

`WORKSPACE`-based documentation is still rendered the same way, and will
not be changed, to keep compatibility. at some point, once we have
modules, we will remove the temporary sub-module `WORKSPACE` files, and
the top level will import their local dependencies as packages again.