Pass the unwrapped `cc` to `wrapCCWith`
The `wrapCCWith` function assumes that the `cc` attribute it is passed is an unwrapped
`cc` since it adds its `/lib` directory to the linker search path.

Passing an already wrapped `cc` here causes `ld` to issue warnings about the directory was
not found, since for a wrapped `cc` it does not exist.