Add `-resource-dir` flag to nix-support/cc-cflags on Darwin
This fixes a problem on Darwin when running `bazel coverage` failed because the
runtime profiling library (libclang_rt.profile_osx.a) was not found.

(see https://github.com/tweag/rules_haskell/issues/1768)

Note, this sets the resource dir to the `/resource-root` directory of the
original wrapped `cc`; otherwise we would have to re-create this directory for
our own wrapper again.