Avoid restarting nixpkgs_package
TL;DR: pre-resolve all repository_ctx.path() before cp() calls.

See bazelbuild/bazel#4533 for background, repository rules in bazel will
be restarted if a depedency it requests is missing.

In the context of rules_nixpkgs, this can cause `nixpkgs_package` to be
restarted every time an additional label is referenced in the rule. This
is especially bad if the rule just happened to finish the slow `cp()`
calls and new labels are referred to after that.

As a workaround, we pre-resolve all labels at the top of repository_rule
and run the `cp`s towards to end.

Signed-off-by: Thomas Lam <thomaslam@canva.com>