Bump nix version to 2.7
Nix version 2.4 is deprecated in nixpkgs unstable, and will be
deprecated in the next stable nixpkgs release.
This also adds the nixVersions prefix to the attribute path, because
specifiying a nix version without the nixVersions prefix also
breaks on nixpkgs unstable.
This allows us to more easily test against nixpkgs unstable, which is
necessary for testing against Bazel 6 prerelease at the moment.
This is tested on both nixpkgs-22.05 and nixpkgs-unstable.