Create rules_nixpkgs_core testing module
The goal is to move the tests in rules_nixpkgs_core into this testing
module for BCR compatibility. The tests in rules_nixpkgs_core incur
extra dependencies that should not be part of rules_nixpkgs_core. In
principle, such dependencies can be marked as dev-dependencies.
However, since rules_nixpkgs only consists of repository rules (or in
the future module extensions) those are easier to test in a dedicated
module.