bzlmod compatible location expansion
The labels as expected by the user are passed in as strings in a mapping
to the actual labels which will be exposed with mangled names in the
repository rule implementation. The mapping allows to map the user
defined location variables to the corresponding paths without exposing
mangled labels to the user.