Construct Label in hub-repository
The `Label` to a nixpkgs repository must be constructed within the
correct scope in order to be able to resolve it to canonical form.
This is done by invoking the `Label` constructor within the scope of the
hub-repository itself, which is generated by the same repositories
module extension as the repository workspaces themselves.