Add nixpkgs_http_repository
A more general version of nixpkgs_git_repository that is not limited to
the specific archives offered by Github, but can be applied to other
HTTP provided repositories in general.

This is a pre-requisite for some of the tags for the new Bazel module
extension designed in design/bzlmod/README.md, in particular the http
and github tags.