cc: Add cross_cpu arg to nixpkgs_cc_toolchain_config
Currently, we hardcode the CPU architecture of the toolchain to the host toolchain.
However, this breaks cross compilation as we will probably want to set the CPU
of the toolchain to the target architecture. The `cross_cpu` flag can be used to
specify the target architecture. If the `cross_cpu` flag is not set, we will auto detect the
host CPU architecture and use that for the toolchain.

Is part of an effort to address #347